requirements.txt
daphne==4.1.2
django==5.1.6
djangorestframework==3.15.2
channels==4.2.0
channels-redis==4.2.1
uuid==1.30